// src/app/shared/model/pages.model.ts
export interface apiResultFormat {
  data: [];
  totalData: number;
}
// ...
export interface FeesRecord {
  id?: number;
  feesGroup: string;
  feesType: string;
  feesCode: string;
  dueDate: string;
  amount: number;
  status: string;
  refId: string;
  mode: string;
  datePaid: string;
  discount: string | number;
  fine: number;
  isSelected?:boolean;
}
// ...
export interface feesGroup {
  isSelected: boolean;
  id?: number;
  ids: string;
  feesGroup: string;
  description: string;
  status: string;
}
// ...
export interface feesType {
  isSelected: boolean;
  id?: number;
  ids?: number;
  feesType: string;
  feesCode: string;
  feesGroup: string;
  description: string;
  status: string;
}